+++ tests/haskell/Numeric/LinearAlgebra/DimensionOverflow/Main.hs
@@ -1,0 +1,7 @@
+{-# LANGUAGE NoImplicitPrelude #-}
+
+import Compiler.Num
+import Numeric.LinearAlgebra
+import System.IO (print)
+
+main = print ((2147483647 >< 2) [] :: Matrix Int)