+++ tests/haskell/Numeric/LinearAlgebra/InfiniteInput/Main.hs
@@ -1,0 +1,8 @@
+{-# LANGUAGE NoImplicitPrelude #-}
+
+import Compiler.Enum
+import Compiler.Num
+import Numeric.LinearAlgebra
+import System.IO (print)
+
+main = print (toLists ((2 >< 3) [1..] :: Matrix Int))