+++ tests/haskell/Numeric/LinearAlgebra/ShortInput/Main.hs
@@ -1,0 +1,7 @@
+{-# LANGUAGE NoImplicitPrelude #-}
+
+import Compiler.Num
+import Numeric.LinearAlgebra
+import System.IO (print)
+
+main = print ((2 >< 2) [1, 2, 3] :: Matrix Int)