+++ tests/haskell/Numeric/LinearAlgebra/UnsupportedElement/Main.hs
@@ -1,0 +1,6 @@
+{-# LANGUAGE NoImplicitPrelude #-}
+
+import Numeric.LinearAlgebra
+
+bad :: Matrix Char
+bad = (1 >< 1) ['x']